Clarify role of ApiController attribute (#12023)
