--- aspnetcore/web-api/define-controller/samples/WebApiSample.Api.21/Controllers/MyBaseController.cs
@@ -1,11 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-
-namespace WebApiSample.Api._21.Controllers
-{
-    #region snippet_ControllerSignature
-    [ApiController]
-    public abstract class MyApiControllerBase : ControllerBase
-    {
-    }
-    #endregion
-}
