--- aspnetcore/web-api/define-controller/samples/WebApiSample.Api.21/Controllers/PetsController.cs
@@ -1,59 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using WebApiSample.DataAccess.Models;
-using WebApiSample.DataAccess.Repositories;
-
-namespace WebApiSample.Api._21.Controllers
-{
-    #region snippet_PetsController
-    [Produces("application/json")]
-    [Route("api/[controller]")]
-    public class PetsController : ControllerBase
-    {
-        private readonly PetsRepository _repository;
-
-        public PetsController(PetsRepository repository)
-        {
-            _repository = repository;
-        }
-
-        [HttpGet]
-        public async Task<ActionResult<List<Pet>>> GetAllAsync()
-        {
-            return await _repository.GetPetsAsync();
-        }
-
-        [HttpGet("{id}")]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public async Task<ActionResult<Pet>> GetByIdAsync(int id)
-        {
-            var pet = await _repository.GetPetAsync(id);
-
-            if (pet == null)
-            {
-                return NotFound();
-            }
-
-            return pet;
-        }
-
-        [HttpPost]
-        [ProducesResponseType(StatusCodes.Status201Created)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult<Pet>> CreateAsync(Pet pet)
-        {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            await _repository.AddPetAsync(pet);
-
-            return CreatedAtAction(nameof(GetByIdAsync),
-                new { id = pet.Id }, pet);
-        }
-    }
-    #endregion
-}
