--- aspnetcore/web-api/define-controller/samples/WebApiSample.Api.Pre21/appsettings.json
@@ -1,15 +1,0 @@
-﻿{
-  "Logging": {
-    "IncludeScopes": false,
-    "Debug": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    },
-    "Console": {
-      "LogLevel": {
-        "Default": "Warning"
-      }
-    }
-  }
-}
