--- aspnetcore/web-api/define-controller/samples/WebApiSample.sln
@@ -1,48 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28218.60
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApiSample.Api.21", "WebApiSample.Api.21\WebApiSample.Api.21.csproj", "{4C1FB735-08A8-41B2-8A58-C62A48221C9C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApiSample.Api.Pre21", "WebApiSample.Api.Pre21\WebApiSample.Api.Pre21.csproj", "{EE80B38F-DFDE-43F4-8A37-988884516A89}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApiSample.DataAccess", "WebApiSample.DataAccess\WebApiSample.DataAccess.csproj", "{D15851E9-72AE-455B-9C5D-B572D180E0F1}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{58B44864-9D5E-4645-8369-0D0CC9C00695}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApiSample.Api.22", "WebApiSample.Api.22\WebApiSample.Api.22.csproj", "{40D8638D-A902-4BDB-8E1E-7481F0F434FA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4C1FB735-08A8-41B2-8A58-C62A48221C9C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C1FB735-08A8-41B2-8A58-C62A48221C9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C1FB735-08A8-41B2-8A58-C62A48221C9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C1FB735-08A8-41B2-8A58-C62A48221C9C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EE80B38F-DFDE-43F4-8A37-988884516A89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE80B38F-DFDE-43F4-8A37-988884516A89}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE80B38F-DFDE-43F4-8A37-988884516A89}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE80B38F-DFDE-43F4-8A37-988884516A89}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D15851E9-72AE-455B-9C5D-B572D180E0F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D15851E9-72AE-455B-9C5D-B572D180E0F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D15851E9-72AE-455B-9C5D-B572D180E0F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D15851E9-72AE-455B-9C5D-B572D180E0F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{40D8638D-A902-4BDB-8E1E-7481F0F434FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40D8638D-A902-4BDB-8E1E-7481F0F434FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{40D8638D-A902-4BDB-8E1E-7481F0F434FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{40D8638D-A902-4BDB-8E1E-7481F0F434FA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {0935E601-D558-4541-B546-A5CA31E619AC}
-	EndGlobalSection
-EndGlobal
