+++ aspnetcore/web-api/index/samples/2.x/Controllers/MyControllerBase.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace WebApiSample.Controllers
+{
+    #region snippet_MyControllerBase
+    [ApiController]
+    public class MyControllerBase : ControllerBase
+    {
+    }
+    #endregion
+}
