+++ aspnetcore/web-api/index/samples/2.x/DataAccess/Models/Product.cs
