+++ aspnetcore/web-api/index/samples/2.x/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace WebApiSample.Api._22
+namespace WebApiSample
 {
     public class Program
     {
