+++ aspnetcore/web-api/index/samples/2.x/appsettings.Development.json
